# Class5/openapi.yaml
swagger: "2.0"
info:
  title: Books API
  description: Simple API for managing books
  version: "1.0"

basePath: /api/v1

paths:
  /books:
    get:
      summary: Get list of books
      parameters:
        - name: page
          in: query
          type: integer
        - name: limit
          in: query
          type: integer
        - name: q
          in: query
          type: string
      responses:
        200:
          description: OK

    post:
      summary: Create a book
      parameters:
        - in: body
          name: body
          schema:
            type: object
            required:
              - title
              - author
            properties:
              title:
                type: string
              author:
                type: string
      responses:
        201:
          description: Created

  /books/{book_id}:
    get:
      summary: Get book by ID
      parameters:
        - name: book_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: OK

    put:
      summary: Update book
      parameters:
        - name: book_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              title:
                type: string
              author:
                type: string
      responses:
        200:
          description: OK

    delete:
      summary: Delete book
      parameters:
        - name: book_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Deleted

  /users/{user_id}/orders:
    get:
      summary: Get orders of a user with multiple pagination strategies
      description: Supports page-based, offset-limit, and cursor pagination via query params.
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
        - name: strategy
          in: query
          type: string
          enum: [page, offset, cursor]
          description: Choose pagination strategy
        - name: page
          in: query
          type: integer
          description: Used when strategy=page
        - name: offset
          in: query
          type: integer
          description: Used when strategy=offset
        - name: limit
          in: query
          type: integer
          description: Number of items per page/chunk
        - name: cursor
          in: query
          type: string
          description: Last seen order id, used when strategy=cursor
      responses:
        200:
          description: OK
        404:
          description: User not found